gcs_event_translator/requirements.txt
functions-framework==3.*
requests
google-auth
cloudevents==1.11.0